requirements.txt
pynput
Pillow
python-dotenv
requests 